refinement: move s3 into aws namespace.
